Fix issue with negative cost when next route has duplicated bus services
Due to loop buses passing the same spot.
Nodes are assumed to be uniquely identified by BusStopCode and ServiceNo
But in this case, a stop can have 2 services of the same number but different direction (59471)
Solution is to consider service passed into edge instead of dest.service as edge service is always directly from edge discovery and dest can be from existing collided node.